Moved InvalidCypherArg and CypherResult into their own files.
